requirements.txt
Flask~=2.3.2
numpy~=1.23.5
requests~=2.31.0
tensorflow~=2.12.0
tensorflow-hub==0.12.0
Pillow~=9.5.0